refactoring: add inner method.
